Move platform exception tests
