+++ platform-sdk/base-utility/src/testFixtures/java/org/hiero/base/utility/test/fixtures/assertions/ExceptionAssertions.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.platform.test.fixtures;
+package org.hiero.base.utility.test.fixtures.assertions;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertSame;
